refactor: migrated ingesting and stats endpoints
